Create abstract filterable list activity
Refactor shared aspects of destination and event list activities to a shared base.
